Split slash command catalog
