+++ docs/DECISIONS.md
@@ -85,7 +85,7 @@
 - Command palette ranking lives beside `WorkspaceCommandSurface`, with category and keyword metadata carried by each command. SwiftUI and the Playwright harness both consume that metadata so grouping, selected-row navigation, and shortcut/title/keyword matching stay aligned.
 - Command palette mode prefixes are treated as discoverability shortcuts, not separate palettes. Empty palette search remains action-only, `>` scopes to workspace actions, `/` scopes to slash-command templates, and non-empty bare queries include both actions and slash commands. Selecting a slash entry inserts the command skeleton into the composer and refocuses the composer instead of executing an incomplete command.
 - Slash commands should route to the same workspace actions as visible UI and the command palette whenever possible. `/terminal`, `/browser`, `/worktrees`, `/pr`, and `/env [name]` now share model/tool paths instead of becoming a separate command system.
-- Slash command discovery uses a shared catalog for `/help` and composer suggestions. The catalog suggests only commands the workspace model can execute or prepare today; transient presentation-only actions such as active Find stay in visible buttons, shortcuts, and the command palette until there is a model-to-view presentation request channel. Suggestions are keyboard-first: arrow keys move the active row, Enter accepts incomplete suggestions while exact commands still submit, and Tab always accepts the active suggestion.
+- Slash command discovery uses `SlashCommandCatalog.swift` for `/help`, command-palette templates, and composer suggestions, while `SlashCommand.swift` owns parser control flow and structured tool-call construction. The catalog suggests only commands the workspace model can execute or prepare today; transient presentation-only actions such as active Find stay in visible buttons, shortcuts, and the command palette until there is a model-to-view presentation request channel. Suggestions are keyboard-first: arrow keys move the active row, Enter accepts incomplete suggestions while exact commands still submit, and Tab always accepts the active suggestion.
 - Tool results produce two different views: the tool card keeps the full structured raw output, while the assistant chat bubble contains a human final answer for common actions and truncates very long output. This avoids the QuillConnect failure mode where users see raw JSON/cards but no clear answer.
 - Tool cards carry an explicit shared density contract: queued and running cards use `peek`, successful completed cards use `collapsed`, and failed or safety-review cards use `expanded`. Renderers may still keep raw JSON in the DOM/surface for copy and diagnostics, but the first visual read should prioritize the conversation and task result over transport data.
 - Agent runs publish incremental thread snapshots through an async progress callback. The workspace model consumes those snapshots immediately so user turns, live TrustedRouter streaming status, safe assistant drafts, queued tools, running tools, safety blocks, and final results are visible before the full run returns. TrustedRouter action text is consumed as a stream, and only structured `say` action text is allowed into the transcript before finalization; tool JSON stays hidden until it becomes a normal tool card.
